separated mapper into mapper and ros_mapper
